Pulling the commits from main to cad (#4)
* default and hover cursors (#3)

* Resources Folder Added

Making the default cursor change colour on an interactive object uses Resources.Load() instead of materials. To make this easier, I'll set up the placeholder pngs for the default and interactive cursors

* Cursor Placeholders added

* cursors placed in assets

* 2/3 Cursor states

Box collider switched to 2D, interactable tag added to front door empty child.

Default cursor changes to interactive on hover for the door. Will add the 3rd cursor state once there's art assets to test it out

* FrontDoor update

moved frontdoor to scene layer and renamed it with 'container' as a reminder its an empty box for the cursor interaction

* QQQ

* Added scene template and door functionality

* Created base follow camera

* resized carpet

* Saved scene

* Added Background to Interior1

* Cutout instead of trasnaparent

* Updated scene template and fixed navmesh

* Rug logic

* Adjusted Z-axis layering of sprites

* Imported A* Pathfinding Project

* Basic pathfinding logic

* Added Parents Fixed some texture issues

* New Sprites

* Completed A* implementation

* changed resolution

---------

Co-authored-by: Lotharingia-BUR <135478406+Lotharingia-BUR@users.noreply.github.com>
Co-authored-by: EternalNyaki <124942024+EternalNyaki@users.noreply.github.com>